refactor: split gameplay commands into modules
