+++ tests/test_teach_move_command.py
@@ -8,8 +8,8 @@ sys.path.insert(0, ROOT)
 
 
 def load_cmd_module():
-    path = os.path.join(ROOT, "commands", "command.py")
-    spec = importlib.util.spec_from_file_location("commands.command", path)
+    path = os.path.join(ROOT, "commands", "cmd_learn_evolve.py")
+    spec = importlib.util.spec_from_file_location("commands.cmd_learn_evolve", path)
     mod = importlib.util.module_from_spec(spec)
     sys.modules[spec.name] = mod
     spec.loader.exec_module(mod)
